+++ R/common_network.R
@@ -21,12 +21,8 @@
 #' @seealso [summarize_address_range()]
 #' @export
 common_network <- function(address1, address2) {
-  if (!is_ip_address(address1)) {
-    abort("`address1` must be an ip_address vector")
-  }
-  if (!is_ip_address(address2)) {
-    abort("`address2` must be an ip_address vector")
-  }
+  check_address(address1)
+  check_address(address2)
 
   # vector recycling
   args <- vec_recycle_common(address1, address2)