+++ R/ip_version.R
@@ -17,9 +17,7 @@ NULL
 #' @rdname is_ipv6
 #' @export
 is_ipv4 <- function(x) {
-  if (!(is_ip_address(x) || is_ip_network(x))) {
-    abort("`x` must be an ip_address or ip_network vector")
-  }
+  check_ip(x)
 
   # directly returning field enables assignment
   result <- !field(x, "is_ipv6")
@@ -29,9 +27,7 @@ is_ipv4 <- function(x) {
 #' @rdname is_ipv6
 #' @export
 is_ipv6 <- function(x) {
-  if (!(is_ip_address(x) || is_ip_network(x))) {
-    abort("`x` must be an ip_address or ip_network vector")
-  }
+  check_ip(x)
 
   # directly returning field enables assignment
   result <- field(x, "is_ipv6")
@@ -53,9 +49,6 @@ is_ipv6 <- function(x) {
 #' @seealso [is_ipv4()], [is_ipv6()], [prefix_length()]
 #' @export
 max_prefix_length <- function(x) {
-  if (!(is_ip_address(x) || is_ip_network(x))) {
-    abort("`x` must be an ip_address or ip_network vector")
-  }
-
+  check_ip(x)
   ifelse(is_ipv6(x), 128L, 32L)
 }