+++ R/ipv6_transition.R
@@ -38,69 +38,48 @@ NULL
 #' @rdname ipv6-transition
 #' @export
 is_ipv4_mapped <- function(x) {
-  if (!(is_ip_address(x) || is_ip_network(x))) {
-    abort("`x` must be an ip_address or ip_network vector")
-  }
-
+  check_ip(x)
   wrap_is_ipv4_mapped(x)
 }
 
 #' @rdname ipv6-transition
 #' @export
 is_6to4 <- function(x) {
-  if (!(is_ip_address(x) || is_ip_network(x))) {
-    abort("`x` must be an ip_address or ip_network vector")
-  }
-
+  check_ip(x)
   wrap_is_6to4(x)
 }
 
 #' @rdname ipv6-transition
 #' @export
 is_teredo <- function(x) {
-  if (!(is_ip_address(x) || is_ip_network(x))) {
-    abort("`x` must be an ip_address or ip_network vector")
-  }
-
+  check_ip(x)
   wrap_is_teredo(x)
 }
 
 #' @rdname ipv6-transition
 #' @export
 extract_ipv4_mapped <- function(x) {
-  if (!is_ip_address(x)) {
-    abort("`x` must be an ip_address vector")
-  }
-
+  check_address(x)
   wrap_extract_ipv4_mapped(x)
 }
 
 #' @rdname ipv6-transition
 #' @export
 extract_6to4 <- function(x) {
-  if (!is_ip_address(x)) {
-    abort("`x` must be an ip_address vector")
-  }
-
+  check_address(x)
   wrap_extract_6to4(x)
 }
 
 #' @rdname ipv6-transition
 #' @export
 extract_teredo_server <- function(x) {
-  if (!is_ip_address(x)) {
-    abort("`x` must be an ip_address vector")
-  }
-
+  check_address(x)
   wrap_extract_teredo_server(x)
 }
 
 #' @rdname ipv6-transition
 #' @export
 extract_teredo_client <- function(x) {
-  if (!is_ip_address(x)) {
-    abort("`x` must be an ip_address vector")
-  }
-
+  check_address(x)
   wrap_extract_teredo_client(x)
 }