+++ NEWS.md
@@ -13,10 +13,11 @@ To mark the release of ipaddress 1.0.0, some **breaking changes** were made to i
   * The `exploded` argument of `format.ip_address()`, `format.ip_network()` and `format.ip_interface()`.
   * The `replace` argument of `sample_ipv4()`, `sample_ipv6()` and `sample_network()`.
   
-## New features
+## Minor improvements and fixes
 
 * The `is_ipv6` argument of `netmask()` and `hostmask()` defaults to `NULL`. This option prefers an IPv4 mask, but creates an IPv6 mask as a fallback.
 * Updated the `iana_ipv4` dataset.
+* Error formatting has been improved using {cli}.
 
 
 # ipaddress 0.5.6